✨ feat: Add "See our properties" button with scroll functionality in Hero component
This commit adds a "See our properties" button with a custom class name and onClick functionality. When clicked, the button triggers a console.log() message and scrolls to the recentHomesRef using the scrollIntoView() method with behavior set to "smooth". Additionally, a new container class "see-properties-container" is added to the div element for styling purposes.