✨ feat: Add email validation and offer message to OfferModal
This commit adds two features to the OfferModal component:
- Email validation using the `email-validator` package.
- Default message text that is pre-filled in the message field.
The commit also updates the `package.json` and `yarn.lock` files with the addition of the `email-validator` package.

Note: The commit message assumes that this is a new feature being added to the project. If this is not the case, the commit message may need to be adjusted accordingly.