Feat/email validation (#55)
* ✨ feat: Add checkEmail API route

This commit adds a new route to check if an email is valid and not disposable, as well as parse and validate an offer amount and retrieve information about a home. The API route returns a JSON response with either a success or invalid status depending on the input validation results.

* ✨ feat: Add email validation and offer message to OfferModal

This commit adds two features to the OfferModal component:
- Email validation using the `email-validator` package.
- Default message text that is pre-filled in the message field.
The commit also updates the `package.json` and `yarn.lock` files with the addition of the `email-validator` package.